Fix menu not opening, related to error: `Any symbols to be exported from a module must be defined with 'var'`.
The full error is:
```
Some code accessed the property 'ArgosMenuItem' on the module 'menuitem'. That property was defined with 'let' or 'const' inside the module. This was previously supported, but is not correct according to the ES6 standard. Any symbols to be exported from a module must be defined with 'var'. The property access will work as previously for the time being, but please fix your code anyway.
```

The menu doesn't seem to open without this change.